debug.sh will report incorrect message
When the /etc/hosts has extra spaces in the end of the line or the
node name is not the last token in the line the node is not discovered
and the result is error messages in the report.

This is fixed by using "getent ahostsv4" to look for the host name.

Fixes: 1328009

Signed-off-by: Phil Cameron <pcameron@redhat.com>