IP Failover diagnostic script
As an OpenShift administrator, I want to know when my ipfailover is
causing my network problems.

Extend the nework debug script so that it can determine why keepalived
is not serving an address appropriately. Ideally we work out which
machine the address should be assigned to, then check that only that
machine has the address assigned, and (ideally) ping the address from
other machines to ensure connectivity.

Trello card:
https://trello.com/c/yFco7WvY/268-5-ip-failover-diagnostic-script-demo-supportability-p3-5

Signed-off-by: Phil Cameron <pcameron@redhat.com>